requirements.txt
fastapi[standard]
sqlalchemy
alembic
fastapi-swagger
passlib
pyjwt